Rename models package to metadata
Update docs
