+++ perdoo/metadata/_base.py
