+++ perdoo/metadata/comic_info.py
@@ -10,7 +10,7 @@ from PIL import Image
 from pydantic import HttpUrl, NonNegativeFloat
 from pydantic_xml import attr, computed_attr, element, wrapped
 
-from perdoo.models._base import PascalModel
+from perdoo.metadata._base import PascalModel
 from perdoo.utils import sanitize
