+++ perdoo/services/marvel.py
@@ -13,7 +13,7 @@ from rich.prompt import Confirm, Prompt
 
 from perdoo import get_cache_root
 from perdoo.console import CONSOLE, create_menu
-from perdoo.models import ComicInfo, MetronInfo
+from perdoo.metadata import ComicInfo, MetronInfo
 from perdoo.services._base import BaseService
 from perdoo.settings import Marvel as MarvelSettings
 from perdoo.utils import IssueSearch, Search, SeriesSearch
@@ -119,7 +119,7 @@ class Marvel(BaseService[Series, Comic]):
             return None
 
     def _process_metron_info(self, series: Series, issue: Comic) -> MetronInfo | None:
-        from perdoo.models.metron_info import (
+        from perdoo.metadata.metron_info import (
             GTIN,
             AgeRating,
             Arc,
@@ -185,7 +185,7 @@ class Marvel(BaseService[Series, Comic]):
         )
 
     def _process_comic_info(self, series: Series, issue: Comic) -> ComicInfo | None:
-        from perdoo.models.comic_info import AgeRating
+        from perdoo.metadata.comic_info import AgeRating
 
         def load_age_rating(value: str) -> AgeRating:
             try:
